+++ cmd/adriatic-map/main.go
@@ -19,7 +19,7 @@ import (
 	"github.com/luxzg/adriatic-map/internal/server"
 )
 
-const version = "0.1.1"
+const version = "0.1.2"
 
 func main() {
 	listenAddress := flag.String("listen", "127.0.0.1:8080", "HTTP listen address")